Renames GetCategoryStatus to GetCategoryIsEnabled
