+++ src/PrestaShopBundle/Resources/config/services/adapter/category.yml
@@ -110,7 +110,7 @@ services:
         command: PrestaShop\PrestaShop\Core\Domain\Category\Command\UpdateCategoryPositionCommand
 
   prestashop.adapter.category.command_handler.get_category_status_handler:
-    class: PrestaShop\PrestaShop\Adapter\Category\QueryHandler\GetCategoryStatusHandler
+    class: PrestaShop\PrestaShop\Adapter\Category\QueryHandler\GetCategoryIsEnabledHandler
     tags:
       - name: tactician.handler
-        command: PrestaShop\PrestaShop\Core\Domain\Category\Query\GetCategoryStatus
+        command: PrestaShop\PrestaShop\Core\Domain\Category\Query\GetCategoryIsEnabled
